Reorganize code.
Split recordServer and replayServer into two new files http_replayer.go and http_recorder.go. Also rename RecordReplayServer to just Server.
