+++ pkg/playback/interactions_test.go
